test(core): split legend layout suites
